font: move all fonts handling to src/font/
