+++ tests/test_bbulk.c
@@ -10,7 +10,7 @@
 #include "text.h"
 
 /* ---- Stubs for external dependencies used by text_bbulk.c ---- */
-#include "../src/font.h"	/* for kmscon_font_* */
+#include "../src/font/font.h"	/* for kmscon_font_* */
 #include "../src/uterm_video.h" /* for uterm_display_* prototypes */
 unsigned int uterm_display_get_width(struct uterm_display *disp)
 {
